+++ src/Facades/TaxOut.php
@@ -16,6 +16,20 @@ class TaxOut
             'ar_invoice_id' => $invoice->id,
             'branch_id' => $invoice->branch_id,
             'chart_of_account_title_id' => $invoice->chart_of_account_title_id,
+
+            'customer_id' => $invoice->payer,
+            'invoice_date' => $invoice->date,
+            'customer_name' => $invoice->payer_name,
+            'invoice_uuid' => $invoice->uuid,
+            'invoice_number' => $invoice->number,
+            'customer_uuid' => $invoice->payer_ref->uuid,
+            'cid' => $invoice->payer_cid,
+            'invoice_ppn' => $invoice->tax,
+            'invoice_dpp' => $invoice->tax_base,
+            'npwp' => $invoice->payer_ref->npwp,
+            'brand_id' => $invoice->brand_id,
+            'customer_address' => $invoice->payer_ref->address,
+            'faktur_id' => $invoice->faktur_id,
         ]);
 
         return $tax_out;
@@ -27,9 +41,7 @@ class TaxOut
         $log->save('debug');
 
         $tax_out = $invoice->tax_out;
-        if ($tax_out) {
-            $tax_out->delete();
-        }
+        if ($tax_out) $tax_out->delete();
 
         return true;
     }
